Reorganize test directory
